Revert "Feature/view endorsement requirements"
