--- data/seeds/001_add_tracks.js
@@ -1,9 +1,0 @@
-exports.seed = function(knex) {
-    return knex("tracks").insert([
-        { title: "Full-Stack Web" },
-        { title: "iOS" },
-        { title: "Data Science" },
-        { title: "Android" },
-        { title: "UX Design" }
-    ])
-}
