--- data/seeds/007_add_user_steps_completed_table.js
@@ -1,7 +1,0 @@
-exports.seed = function(knex) {
-    return knex("user_steps_completed").insert([
-        { user_id: 1, steps_id: 1 },
-        { user_id: 1, steps_id: 2 },
-        { user_id: 1, steps_id: 6 }
-    ])
-}
