+++ package.json
@@ -12,7 +12,7 @@
         "knex:rollback:all": "knex migrate:rollback --all",
         "knex:seed:run": "knex seed:run",
         "start": "node -r dotenv/config .",
-        "test": "(dropdb test_endrsd && createdb test_endrsd) && jest --runInBand"
+        "test": "(dropdb test_career_endorsement && createdb test_career_endorsement) && jest --runInBand"
     },
     "repository": "git@github.com:labs15-career-endorsement-tracker/backend.git",
     "author": "Labs15SuperTeam",
@@ -27,15 +27,12 @@
         "supertest": "^4.0.2"
     },
     "dependencies": {
-        "bcryptjs": "^2.4.3",
         "body-parser": "^1.19.0",
         "cors": "^2.8.5",
         "dotenv": "^8.0.0",
         "express": "^4.17.1",
-        "express-validator": "^6.1.1",
         "helmet": "^3.20.0",
         "http-errors": "^1.7.3",
-        "jsonwebtoken": "^8.5.1",
         "knex": "^0.19.1",
         "morgan": "^1.9.1",
         "pg": "^7.12.1"
