+++ src/config.js
@@ -5,23 +5,11 @@ require("dotenv").config({
     path: resolvePath(__dirname, "../.env")
 })
 
-const {
-    NODE_ENV,
-    DATABASE_URL,
-    TEST_DATABASE_URL,
-    PORT,
-    JWT_SECRET
-} = process.env
+const { NODE_ENV, DATABASE_URL, TEST_DATABASE_URL, PORT } = process.env
 
 const environment = NODE_ENV || "development"
 const isTesting = environment === "test"
 const dbUrl = isTesting ? TEST_DATABASE_URL : DATABASE_URL
 const port = PORT || 5000
 
-module.exports = {
-    environment,
-    dbUrl,
-    port,
-    version: version.split(".")[0],
-    jwtSecret: JWT_SECRET
-}
+module.exports = { environment, dbUrl, port, version: version.split(".")[0] }
