+++ src/controllers/index.js
@@ -10,8 +10,4 @@ const getUsers = async (req, res, next) => {
     }
 }
 
-module.exports = {
-    getUsers,
-    login: require("./login"),
-    viewEndorsementRequirements: require("./viewEndorsementRequirements")
-}
+module.exports = { getUsers }
