+++ src/middleware/index.js
@@ -1,11 +1,16 @@
-const { handle404, handle500 } = require("./error-handlers")
-const { validateEmail, validatePassword } = require("./validators")
-const requiresAuth = require("./requiresAuth")
+const { NotFound } = require("http-errors")
+const handle404 = (req, _res, next) => {
+    const { method, path } = req
+    const msg = `${method} ${path} has not been implemented.`
+    next(NotFound(msg))
+}
+
+const handle500 = ({ status = 500, name, message }, _req, res, next) => {
+    if (res.headersSent) return next()
+    res.status(status).json({ name, statusCode: status, message })
+}
 
 module.exports = {
     handle404,
-    handle500,
-    validateEmail,
-    validatePassword,
-    requiresAuth
+    handle500
 }
