+++ src/model/index.js
@@ -1,18 +1,5 @@
 const db = require("../../data")
-const {
-    findRequirementsByTrack,
-    getRequirementsWithProgress
-} = require("./requirements")
-
-const { findCompletedStepsBy, findStepsByTask } = require("./steps")
 
 const findUsers = () => db("users")
 
-const findUsersBy = filter => db("users").where(filter)
-
-module.exports = {
-    findUsers,
-    findUsersBy,
-    findRequirementsByTrack,
-    getRequirementsWithProgress
-}
+module.exports = { findUsers }
