+++ src/routes/index.js
@@ -1,19 +1,9 @@
 const { Router } = require("express")
 
-const {
-    getUsers,
-    login,
-    viewEndorsementRequirements
-} = require("../controllers")
-
-const { requiresAuth } = require("../middleware")
+const { getUsers } = require("../controllers")
 
 const api = Router()
 
 api.get("/users", getUsers)
-
-api.post("/login", login)
-
-api.get("/view-requirements", requiresAuth, viewEndorsementRequirements)
 
 module.exports = api
